# Parse \uXXXX escapes faster