@purplesyringa 2024-08-11
Is changing error precision okay? I couldn't sidestep that without sacrificing performance.