@dtolnay 2024-08-11
> Is changing error precision (see the changed test) okay? I couldn't sidestep that without sacrificing performance.

The code previously pointed to a specific one of the expected hex digits:

```haskell
"\u0000\u00#0\u0000"
           ^
```

Preserving that is not important to me.

The new code always points to the 6th byte after the backslash.

```haskell
"\u0000\u00#0\u0000"
             ^
```

Would it be costly to point to the current escape's backslash instead? This would make more sense and hopefully be as simple as offsetting the index by -6 in the error codepath.

```haskell
"\u0000\u00#0\u0000"
       ^
```