@purplesyringa 2024-08-11
I can do that easily in the `str` case but not the I/O case. Is that fine?