@purplesyringa 2024-08-11
Hmm. Now that I think about it, I'm not sure if I understand how error location works. `position()` is typically called after a call to `next()`, so the error points at the byte *after* the wrong one.

So e.g. here serde-json reports the error immediately after consuming malformed UTF-8:
https://github.com/serde-rs/json/blob/cf771a0471dd797b6fead77e767f2f7943740c98/tests/test.rs#L1080-L1083

And here we say "line 2" after consuming the erroneous `\n`:
https://github.com/serde-rs/json/blob/cf771a0471dd797b6fead77e767f2f7943740c98/tests/test.rs#L1108-L1111

I think throwing up pointing after an invalid escape (i.e. the current behavior) would be more consistent.