@dtolnay 2024-08-12
> I'm not sure if I understand how error location works. `position()` is typically called after a call to `next()`, so the error points at the byte *after* the wrong one.

Ideally the position returned by `position()` is supposed to point at the byte most recently produced by `next()`. Separately there is `peek_position()` to return the position of the next byte not yet returned by `next()`.

Sometimes there are codepaths where the byte which caused an error might have come from either `next()` or `peek()`, so the error position won't always be right in that case. Sometimes it's possible to move the error position computation earlier, so that it knows whether `position()` or `peek_position()` is the one to use. It is not intended in general that the reported error positions always refer to the next byte after the bad one.

https://github.com/serde-rs/json/blob/b4bc6436acdf874729a7b425b3c2e48f8b16b398/src/read.rs#L39-L57