@dtolnay 2024-08-12
> I can do that easily in the `str` case but not the I/O case. Is that fine?

That would be fine. Something approximate would also be fine, like `column.saturating_sub(6)` in the I/O case, which will be correct almost all of the time with the exception of something grossly malformed like:

```
"\u00
0"
```